backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.13
django-cors-headers==4.4.0
django-elasticsearch-dsl==8.0
django-nine==0.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
elastic-transport==8.13.1
elasticsearch==8.14.0
elasticsearch-dsl==8.14.0
idna==3.7
packaging==24.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
stripe==10.2.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
